fix(converter.py): Bifurcated ros2 msg and ulog msg
There was an issue where the ros2 msg name wasn't matching ulog msg name
because we wanted to go with camel case over snake case. This is caught
and fixed by separating them into 2 strings.

There is probably a better solution to this, but it works at the moment
without issue.

Might need to test edge cases.